[cmd/telemetrygen] move common package into config
Updating the common package name to config and move the code into the right dir. Part of appealing golangci-lint
