+++ cmd/telemetrygen/internal/config/config.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package common
+package config
 
 import (
 	"errors"
