+++ cmd/telemetrygen/pkg/logs/worker.go
@@ -20,7 +20,7 @@ import (
 	"go.uber.org/zap"
 	"golang.org/x/time/rate"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/cmd/telemetrygen/internal/common"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/cmd/telemetrygen/internal/config"
 	types "github.com/open-telemetry/opentelemetry-collector-contrib/cmd/telemetrygen/pkg"
 )
 
@@ -68,7 +68,7 @@ func (w worker) simulateLogs(res *resource.Resource, exporter sdklog.Exporter, t
 		// Add load size attributes if specified
 		if w.loadSize > 0 {
 			for j := 0; j < w.loadSize; j++ {
-				attrs = append(attrs, log.String(fmt.Sprintf("load-%v", j), string(make([]byte, common.CharactersPerMB))))
+				attrs = append(attrs, log.String(fmt.Sprintf("load-%v", j), string(make([]byte, config.CharactersPerMB))))
 			}
 		}
