+++ cmd/telemetrygen/pkg/logs/worker_test.go
@@ -14,7 +14,7 @@ import (
 	sdklog "go.opentelemetry.io/otel/sdk/log"
 	"go.uber.org/zap"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/cmd/telemetrygen/internal/common"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/cmd/telemetrygen/internal/config"
 	types "github.com/open-telemetry/opentelemetry-collector-contrib/cmd/telemetrygen/pkg"
 )
 
@@ -48,7 +48,7 @@ func (*mockExporter) ForceFlush(context.Context) error {
 
 func TestFixedNumberOfLogs(t *testing.T) {
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount: 1,
 		},
 		NumLogs:        5,
@@ -73,7 +73,7 @@ func TestFixedNumberOfLogs(t *testing.T) {
 
 func TestDurationInf(t *testing.T) {
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			TotalDuration: types.DurationWithInf(-1),
 		},
 		SeverityText:   "Info",
@@ -90,7 +90,7 @@ func TestDurationInf(t *testing.T) {
 
 func TestRateOfLogs(t *testing.T) {
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			Rate:          10,
 			TotalDuration: types.DurationWithInf(time.Second / 2),
 			WorkerCount:   1,
@@ -115,7 +115,7 @@ func TestRateOfLogs(t *testing.T) {
 
 func TestUnthrottled(t *testing.T) {
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			TotalDuration: types.DurationWithInf(1 * time.Second),
 			WorkerCount:   1,
 		},
@@ -138,7 +138,7 @@ func TestCustomBody(t *testing.T) {
 	cfg := &Config{
 		Body:    "custom body",
 		NumLogs: 1,
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount: 1,
 		},
 		SeverityText:   "Info",
@@ -275,7 +275,7 @@ func TestValidate(t *testing.T) {
 		{
 			name: "No duration, NumLogs",
 			cfg: &Config{
-				Config: common.Config{
+				Config: config.Config{
 					WorkerCount: 1,
 				},
 				TraceID: "123",
@@ -285,7 +285,7 @@ func TestValidate(t *testing.T) {
 		{
 			name: "TraceID invalid",
 			cfg: &Config{
-				Config: common.Config{
+				Config: config.Config{
 					WorkerCount: 1,
 				},
 				NumLogs: 5,
@@ -296,7 +296,7 @@ func TestValidate(t *testing.T) {
 		{
 			name: "SpanID invalid",
 			cfg: &Config{
-				Config: common.Config{
+				Config: config.Config{
 					WorkerCount: 1,
 				},
 				NumLogs: 5,
@@ -308,7 +308,7 @@ func TestValidate(t *testing.T) {
 		{
 			name: "LoadSize negative",
 			cfg: &Config{
-				Config: common.Config{
+				Config: config.Config{
 					WorkerCount: 1,
 					LoadSize:    -1,
 				},
@@ -333,7 +333,7 @@ func configWithNoAttributes(qty int, body string) *Config {
 	return &Config{
 		Body:    body,
 		NumLogs: qty,
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount:         1,
 			TelemetryAttributes: nil,
 		},
@@ -346,9 +346,9 @@ func configWithOneAttribute(qty int, body string) *Config {
 	return &Config{
 		Body:    body,
 		NumLogs: qty,
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount:         1,
-			TelemetryAttributes: common.KeyValue{telemetryAttrKeyOne: telemetryAttrValueOne},
+			TelemetryAttributes: config.KeyValue{telemetryAttrKeyOne: telemetryAttrValueOne},
 		},
 		SeverityText:   "Info",
 		SeverityNumber: 9,
@@ -356,7 +356,7 @@ func configWithOneAttribute(qty int, body string) *Config {
 }
 
 func configWithMultipleAttributes(qty int, body string) *Config {
-	kvs := common.KeyValue{
+	kvs := config.KeyValue{
 		telemetryAttrKeyOne:     telemetryAttrValueOne,
 		telemetryAttrKeyTwo:     telemetryAttrValueTwo,
 		telemetryAttrIntKeyOne:  telemetryAttrIntValueOne,
@@ -365,7 +365,7 @@ func configWithMultipleAttributes(qty int, body string) *Config {
 	return &Config{
 		Body:    body,
 		NumLogs: qty,
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount:         1,
 			TelemetryAttributes: kvs,
 		},
@@ -377,7 +377,7 @@ func configWithMultipleAttributes(qty int, body string) *Config {
 func TestLogsWithLoadSize(t *testing.T) {
 	// arrange
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount: 1,
 			LoadSize:    2, // 2MB of load data
 		},
@@ -409,11 +409,11 @@ func TestLogsWithLoadSize(t *testing.T) {
 	logRecord.WalkAttributes(func(attr log.KeyValue) bool {
 		if attr.Key == "load-0" {
 			load0Found = true
-			assert.Len(t, attr.Value.AsString(), common.CharactersPerMB, "load-0 should have 1MB of data")
+			assert.Len(t, attr.Value.AsString(), config.CharactersPerMB, "load-0 should have 1MB of data")
 		}
 		if attr.Key == "load-1" {
 			load1Found = true
-			assert.Len(t, attr.Value.AsString(), common.CharactersPerMB, "load-1 should have 1MB of data")
+			assert.Len(t, attr.Value.AsString(), config.CharactersPerMB, "load-1 should have 1MB of data")
 		}
 		return true
 	})
@@ -451,7 +451,7 @@ func TestLogsWithDefaultLoadSize(t *testing.T) {
 func TestLogsWithZeroLoadSize(t *testing.T) {
 	// arrange
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount: 1,
 			LoadSize:    0, // Disable load size
 		},
