+++ cmd/telemetrygen/pkg/metrics/worker_test.go
@@ -19,7 +19,7 @@ import (
 	"go.uber.org/zap"
 	"golang.org/x/time/rate"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/cmd/telemetrygen/internal/common"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/cmd/telemetrygen/internal/config"
 	types "github.com/open-telemetry/opentelemetry-collector-contrib/cmd/telemetrygen/pkg"
 )
 
@@ -77,7 +77,7 @@ func checkMetricTemporality(t *testing.T, ms metricdata.Metrics, metricType Metr
 func TestFixedNumberOfMetrics(t *testing.T) {
 	// arrange
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount: 1,
 		},
 		NumMetrics: 5,
@@ -99,7 +99,7 @@ func TestFixedNumberOfMetrics(t *testing.T) {
 
 func TestDurationInf(t *testing.T) {
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			TotalDuration: types.DurationWithInf(-1),
 		},
 		MetricType: MetricTypeSum,
@@ -116,7 +116,7 @@ func TestDurationInf(t *testing.T) {
 func TestRateOfMetrics(t *testing.T) {
 	// arrange
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			Rate:          10,
 			TotalDuration: types.DurationWithInf(time.Second / 2),
 			WorkerCount:   1,
@@ -175,7 +175,7 @@ func TestMetricsWithTemporality(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			// arrange
 			cfg := &Config{
-				Config: common.Config{
+				Config: config.Config{
 					WorkerCount: 1,
 				},
 				NumMetrics:             1,
@@ -207,7 +207,7 @@ func TestMetricsWithTemporality(t *testing.T) {
 func TestUnthrottled(t *testing.T) {
 	// arrange
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			TotalDuration: types.DurationWithInf(1 * time.Second),
 			WorkerCount:   1,
 		},
@@ -423,7 +423,7 @@ func TestValidate(t *testing.T) {
 		{
 			name: "No duration, NumMetrics, or Continuous",
 			cfg: &Config{
-				Config: common.Config{
+				Config: config.Config{
 					WorkerCount: 1,
 				},
 				MetricType: MetricTypeSum,
@@ -434,7 +434,7 @@ func TestValidate(t *testing.T) {
 		{
 			name: "TraceID invalid",
 			cfg: &Config{
-				Config: common.Config{
+				Config: config.Config{
 					WorkerCount: 1,
 				},
 				NumMetrics: 5,
@@ -446,7 +446,7 @@ func TestValidate(t *testing.T) {
 		{
 			name: "SpanID invalid",
 			cfg: &Config{
-				Config: common.Config{
+				Config: config.Config{
 					WorkerCount: 1,
 				},
 				NumMetrics: 5,
@@ -459,7 +459,7 @@ func TestValidate(t *testing.T) {
 		{
 			name: "LoadSize negative",
 			cfg: &Config{
-				Config: common.Config{
+				Config: config.Config{
 					WorkerCount: 1,
 					LoadSize:    -1,
 				},
@@ -483,7 +483,7 @@ func TestValidate(t *testing.T) {
 
 func configWithNoAttributes(metric MetricType, qty int) *Config {
 	return &Config{
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount:         1,
 			TelemetryAttributes: nil,
 		},
@@ -495,9 +495,9 @@ func configWithNoAttributes(metric MetricType, qty int) *Config {
 
 func configWithOneAttribute(metric MetricType, qty int) *Config {
 	return &Config{
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount:         1,
-			TelemetryAttributes: common.KeyValue{telemetryAttrKeyOne: telemetryAttrValueOne},
+			TelemetryAttributes: config.KeyValue{telemetryAttrKeyOne: telemetryAttrValueOne},
 		},
 		NumMetrics: qty,
 		MetricName: "test",
@@ -506,14 +506,14 @@ func configWithOneAttribute(metric MetricType, qty int) *Config {
 }
 
 func configWithMultipleAttributes(metric MetricType, qty int) *Config {
-	kvs := common.KeyValue{
+	kvs := config.KeyValue{
 		telemetryAttrKeyOne:     telemetryAttrValueOne,
 		telemetryAttrKeyTwo:     telemetryAttrValueTwo,
 		telemetryAttrIntKeyOne:  telemetryAttrIntValueOne,
 		telemetryAttrBoolKeyOne: telemetryAttrBoolValueOne,
 	}
 	return &Config{
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount:         1,
 			TelemetryAttributes: kvs,
 		},
@@ -524,7 +524,7 @@ func configWithMultipleAttributes(metric MetricType, qty int) *Config {
 
 func configWithEnabledUnique(metric MetricType, qty int) *Config {
 	return &Config{
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount:         1,
 			TelemetryAttributes: nil,
 		},
@@ -650,7 +650,7 @@ func TestUniqueSumTimeseries(t *testing.T) {
 func TestMetricsWithLoadSize(t *testing.T) {
 	// arrange
 	cfg := &Config{
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount: 1,
 			LoadSize:    2, // 2MB of load data
 		},
@@ -680,11 +680,11 @@ func TestMetricsWithLoadSize(t *testing.T) {
 	// Check that load attributes exist and have the expected size
 	load0Value, exists := attr.Value("load-0")
 	assert.True(t, exists, "should have load-0 attribute")
-	assert.Len(t, load0Value.AsString(), common.CharactersPerMB, "load-0 should have 1MB of data")
+	assert.Len(t, load0Value.AsString(), config.CharactersPerMB, "load-0 should have 1MB of data")
 
 	load1Value, exists := attr.Value("load-1")
 	assert.True(t, exists, "should have load-1 attribute")
-	assert.Len(t, load1Value.AsString(), common.CharactersPerMB, "load-1 should have 1MB of data")
+	assert.Len(t, load1Value.AsString(), config.CharactersPerMB, "load-1 should have 1MB of data")
 }
 
 func TestMetricsWithDefaultLoadSize(t *testing.T) {
@@ -720,7 +720,7 @@ func TestExponentialHistogramMetricGeneration(t *testing.T) {
 	// arrange
 	m := &mockExporter{}
 	cfg := Config{
-		Config: common.Config{
+		Config: config.Config{
 			WorkerCount: 1,
 		},
 		NumMetrics: 1,
